297:fix/removing useEffect from useAuth, adding the useEffect into _app.tsx
